support deleteMulti
