+++ package.json
@@ -45,6 +45,7 @@
     "debug": "~2.1.1",
     "humanize-ms": "~1.0.1",
     "is-type-of": "~0.3.1",
+    "merge-descriptors": "~0.0.2",
     "mime": "~1.3.4",
     "urllib": "~2.3.0",
     "utility": "~1.3.0",
